+++ CHANGELOG.md
@@ -3,7 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
-## [0.4.12] - 2022-05.12
+## [0.4.13] - 2022-08-25
+### Added
+ARGUS-29459
+- Added `LocalLoggingContext` to simplify setting a specific logging context for a specified block of code.
+- Added `DeprecatedLoggingContext` to simplify adding a "deprecated" logging context variable for a specified block of code.
+
+## [0.4.12] - 2022-05-12
 ### Changed
 ARGUS-28992
 - Exposing all context variables in `LoggingContext.getAll()`